Remove assumptions that libc::c_char and core::ffi::c_char are the same type
When cross-compiling for PS Vita from Apple Silicon (aarch64), core::ffi:c_char
and libc::c_char have different signed-ness (i8 vs u8). I think this might be
a bug in cargo-vita rather than in SDL2, however, the easiest quick-fix is to
remove all assumptions in this binding code that they are the same, by casting
to *const _, which lets the compiler pick the right type.